requirements.txt
# Brain Tumor Detection System - Requirements
# Core dependencies for Streamlit cloud-based application

# Web Framework
streamlit==1.31.1
streamlit-option-menu==0.3.12

# Machine Learning & Deep Learning
tensorflow==2.14.0
keras==2.14.0
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.3
scipy==1.11.4

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Cloud & Database
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development (optional)
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0
